correção validação COR
