+++ src/main/java/domain/pagamentos/services/strategies/EstrategiaPagamentoCartaoCredito.java
@@ -22,13 +22,8 @@ public class EstrategiaPagamentoCartaoCredito implements EstrategiaPagamento {
         System.out.println("Quantidade de parcelas (mínimo 1): ");
         dados.setQuantidadeParcelas(scanner.nextInt());
         pagamento.setDadosPagamento(dados);
-        System.out.printf("Parcelas serão de R$ %.2f", calcularParcelas(pagamento));
+
+
     }
 
-    private BigDecimal calcularParcelas(Pagamento pagamento){
-        int quantidadeParcelas = ((DadosCartaoCredito) pagamento.getDadosPagamento()).getQuantidadeParcelas();
-        BigDecimal valorTotal = pagamento.getValor();
-
-        return valorTotal.divide(new BigDecimal(quantidadeParcelas), RoundingMode.HALF_UP);
-    }
 }
